+++ dddev/backend/src/main/java/com/d103/dddev/api/ground/service/GroundService.java
@@ -19,9 +19,9 @@ public interface GroundService {
 	Map<String, Integer> getGroundActiveTime(Integer groundId, Integer sprintId) throws Exception;
 	Map<String, Integer> getGroundTotalTime(Integer groundId, Integer sprintId) throws Exception;
 
-	Map<String, Long> getGroundFocusTimeCount(Integer groundId, Integer sprintId) throws Exception;
-	Map<String, Long> getGroundActiveTimeCount(Integer groundId, Integer sprintId) throws Exception;
-	Map<String, Long> getGroundTotalTimeCount(Integer groundId, Integer sprintId) throws Exception;
+	Map<String, Integer> getGroundFocusTimeCount(Integer groundId, Integer sprintId) throws Exception;
+	Map<String, Integer> getGroundActiveTimeCount(Integer groundId, Integer sprintId) throws Exception;
+	Map<String, Integer> getGroundTotalTimeCount(Integer groundId, Integer sprintId) throws Exception;
 
 	GroundDto updateGroundInfo(GroundDto newGroundDto, GroundDto groundDto) throws Exception;
 	GroundDto updateGroundProfile(MultipartFile file, GroundDto groundDto) throws Exception;
