+++ dddev/backend/src/main/java/com/d103/dddev/api/issue/service/IssueService.java
@@ -32,7 +32,7 @@ public interface IssueService  {
     Map<String, Integer> getGroundActiveTime(Integer groundId, Integer sprintId) throws Exception;
     Map<String, Integer> getGroundTotalTime(Integer groundId, Integer sprintId) throws Exception;
 
-    Map<String, Long> getGroundFocusTimeCount(Integer groundId, Integer sprintId) throws Exception;
-    Map<String, Long> getGroundActiveTimeCount(Integer groundId, Integer sprintId) throws Exception;
-    Map<String, Long> getGroundTotalTimeCount(Integer groundId, Integer sprintId) throws Exception;
+    Map<String, Integer> getGroundFocusTimeCount(Integer groundId, Integer sprintId) throws Exception;
+    Map<String, Integer> getGroundActiveTimeCount(Integer groundId, Integer sprintId) throws Exception;
+    Map<String, Integer> getGroundTotalTimeCount(Integer groundId, Integer sprintId) throws Exception;
 }
